refactored to use dependency injection
