+++ src/Ucommerce.Masterclass.Umbraco/Models/LoginRequestViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace Ucommerce.Masterclass.Umbraco.Models
+{
+    public class LoginRequestViewModel
+    {
+        public string Username { get; set; }
+
+        public string Password { get; set; }
+    }
+}
