+++ src/Ucommerce.Masterclass.Umbraco/Models/LoginViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace Ucommerce.Masterclass.Umbraco.Models
+{
+    public class LoginViewModel
+    {
+        public bool IsLoggedIn { get; set; }
+
+        public string UserName { get; set; }
+    }
+}
