+++ src/Ucommerce.Masterclass.Umbraco/Views/CategoryNavigation/Index.cshtml
@@ -1,6 +1,7 @@
 ﻿@using Ucommerce.Masterclass.Umbraco.Models
 @model Ucommerce.Masterclass.Umbraco.Models.CategoryNavigationViewModel
 @{
+    Layout = null;
 }
 
 <ul class="list-group list-group-item card bg-light">
